+++ src/main/java/com/mobble/mobbleserver/domain/meeting/Meeting.java
@@ -1,7 +1,7 @@
 package com.mobble.mobbleserver.domain.meeting;
 
 import com.mobble.mobbleserver.domain.clubMember.ClubMember;
-import com.mobble.mobbleserver.domain.common.exception.DomainException;
+import com.mobble.mobbleserver.domain.exception.DomainException;
 import com.mobble.mobbleserver.domain.meeting.error.MeetingError;
 import com.mobble.mobbleserver.domain.member.Member;
 import jakarta.persistence.*;
