+++ src/main/java/com/mobble/mobbleserver/domain/meeting/MeetingSchedule.java
@@ -1,6 +1,6 @@
 package com.mobble.mobbleserver.domain.meeting;
 
-import com.mobble.mobbleserver.domain.common.exception.DomainException;
+import com.mobble.mobbleserver.domain.exception.DomainException;
 import com.mobble.mobbleserver.domain.meeting.error.MeetingError;
 import jakarta.persistence.Column;
 import jakarta.persistence.Embeddable;
